backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
sqlmodel==0.0.27
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
pydantic-settings==2.12.0
pandas==2.3.3
openpyxl==3.1.5
reportlab==4.4.5
httpx==0.28.1
